requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.2.2
jinja2==3.1.2
python-dotenv==1.0.0
tqdm==4.65.0